Complete main, Fix Nth Best Ending Viterbi Bug, Fix options set (preset)
Completed a version of main that allows of multiple algorithms and
multiple outputs.

Fixed NBest viterbi bug that gave incorrect traceback of ending cell.

Fixed option bug that would consider a default value as variable being
set from commandline.  Also changed so that if a preset is set then it
is allowed to not pass in a value