Remove multiple calls to multiTraceback::finalize
multiTraceback::finalize() was being called in StochHMM.cpp at
print_output, and then called again in each of the print
functions (print_path, print_label, print_gff), with the result that
each path was added to pathAccess twice. After sorting, this meant that
only the first half the paths, including duplicates, was printed when,
for instance, the command

% ./stochhmm -model examples/Dice.hmm -seq examples/Dice_short.fa \
  -stochastic forward -repetitions 10 -label

was given.